quan/requirements.txt
streamlit
pycryptodome
pydantic
PyYAML

